refactor: move code from capabilities command to util file in preparation for tests
